+++ temba/channels/tests.py
@@ -7,6 +7,8 @@ import copy
 import hashlib
 import hmac
 import json
+
+import iso8601
 import pytz
 import six
 import time
@@ -4463,11 +4465,33 @@ class InfobipTest(TembaTest):
                                       uuid='00000000-0000-0000-0000-000000001234')
 
     def test_received(self):
-        data = {'receiver': '2347030767144', 'sender': '2347030767143', 'text': 'Hello World'}
-        encoded_message = urlencode(data)
 
-        callback_url = reverse('handlers.infobip_handler', args=['received', self.channel.uuid]) + "?" + encoded_message
-        response = self.client.get(callback_url)
+        now = timezone.now()
+        two_hour_ago = now - timedelta(hours=2)
+
+        post_data = {
+            "results": [
+                {
+                    "messageId": "817790313235066447",
+                    "from": "2347030767143",
+                    "to": "2347030767144",
+                    "text": "Hello World",
+                    "cleanText": "World",
+                    "keyword": "Hello",
+                    "receivedAt": two_hour_ago.isoformat(),
+                    "smsCount": 1,
+                    "price": {
+                        "pricePerMessage": 0,
+                        "currency": "EUR"
+                    },
+                    "callbackData": "callbackData"
+                }],
+            "messageCount": 1,
+            "pendingMessageCount": 0
+        }
+
+        receive_url = reverse('courier.ib', args=[self.channel.uuid, 'receive'])
+        response = self.client.post(receive_url, json.dumps(post_data), content_type='application/json')
 
         self.assertEqual(200, response.status_code)
 
@@ -4478,16 +4502,36 @@ class InfobipTest(TembaTest):
         self.assertEqual(self.org, msg.org)
         self.assertEqual(self.channel, msg.channel)
         self.assertEqual("Hello World", msg.text)
+        self.assertEqual(two_hour_ago, msg.sent_on)
+        self.assertTrue(now > msg.sent_on)
+
+        Msg.objects.all().delete()
+
+        with patch('iso8601.parse_date') as mock:
+            mock.side_effect = [iso8601.ParseError()]
+
+            response = self.client.post(receive_url, json.dumps(post_data), content_type='application/json')
+
+            self.assertEqual(200, response.status_code)
+
+            # load our message
+            msg = Msg.objects.get()
+            self.assertEqual('+2347030767143', msg.contact.get_urn(TEL_SCHEME).path)
+            self.assertEqual(INCOMING, msg.direction)
+            self.assertEqual(self.org, msg.org)
+            self.assertEqual(self.channel, msg.channel)
+            self.assertEqual("Hello World", msg.text)
+            self.assertTrue(now < msg.sent_on)
 
         # try it with an invalid receiver, should fail as UUID and receiver id are mismatched
-        data['receiver'] = '2347030767145'
-        encoded_message = urlencode(data)
-
-        callback_url = reverse('handlers.infobip_handler', args=['received', self.channel.uuid]) + "?" + encoded_message
-        response = self.client.get(callback_url)
+        post_data['results'][0]['to'] = '2347030767145'
+        response = self.client.post(receive_url, json.dumps(post_data), content_type='application/json')
 
         # should get 404 as the channel wasn't found
         self.assertEqual(404, response.status_code)
+
+        response = self.client.get(receive_url)
+        self.assertEqual(405, response.status_code)
 
     def test_delivered(self):
         contact = self.create_contact("Joe", '+2347030767143')
@@ -4495,25 +4539,38 @@ class InfobipTest(TembaTest):
         msg.external_id = '254021015120766124'
         msg.save(update_fields=('external_id',))
 
-        # mark it as delivered
-        base_body = '<DeliveryReport><message id="254021015120766124" sentdate="2014/02/10 16:12:07" ' \
-                    ' donedate="2014/02/10 16:13:00" status="STATUS" gsmerror="0" price="0.65" /></DeliveryReport>'
-        delivery_url = reverse('handlers.infobip_handler', args=['delivered', self.channel.uuid])
+        post_data = {
+            "results": [
+                {
+                    "messageId": msg.id,
+                    "status": {
+                        "groupName": "DELIVERED"
+                    }
+                }
+            ]
+        }
 
-        # assert our SENT status
-        response = self.client.post(delivery_url, data=base_body.replace('STATUS', 'SENT'), content_type='application/xml')
-        self.assertEqual(200, response.status_code)
-        msg = Msg.objects.get()
-        self.assertEqual(SENT, msg.status)
+        delivery_url = reverse('courier.ib', args=[self.channel.uuid, 'delivered'])
 
         # assert our DELIVERED status
-        response = self.client.post(delivery_url, data=base_body.replace('STATUS', 'DELIVERED'), content_type='application/xml')
+        response = self.client.post(delivery_url, json.dumps(post_data), content_type='application/json')
         self.assertEqual(200, response.status_code)
         msg = Msg.objects.get()
         self.assertEqual(DELIVERED, msg.status)
+        msg.status = SENT
+        msg.save()
 
         # assert our FAILED status
-        response = self.client.post(delivery_url, data=base_body.replace('STATUS', 'NOT_SENT'), content_type='application/xml')
+        post_data['results'][0]['status']['groupName'] = 'REJECTED'
+        response = self.client.post(delivery_url, json.dumps(post_data), content_type='application/json')
+        self.assertEqual(200, response.status_code)
+        msg = Msg.objects.get()
+        self.assertEqual(FAILED, msg.status)
+        msg.status = SENT
+        msg.save()
+
+        post_data['results'][0]['status']['groupName'] = 'UNDELIVERABLE'
+        response = self.client.post(delivery_url, json.dumps(post_data), content_type='application/json')
         self.assertEqual(200, response.status_code)
         msg = Msg.objects.get()
         self.assertEqual(FAILED, msg.status)
@@ -4526,7 +4583,7 @@ class InfobipTest(TembaTest):
             settings.SEND_MESSAGES = True
 
             with patch('requests.post') as mock:
-                mock.return_value = MockResponse(200, json.dumps(dict(messages=[{'status': {'id': 0}, 'messageid': 12}])))
+                mock.return_value = MockResponse(200, json.dumps(dict(messages=[{'status': {'groupId': 1}}])))
 
                 # manually send it off
                 Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
@@ -4535,7 +4592,19 @@ class InfobipTest(TembaTest):
                 msg.refresh_from_db()
                 self.assertEqual(SENT, msg.status)
                 self.assertTrue(msg.sent_on)
-                self.assertEqual('12', msg.external_id)
+                self.assertIsNone(msg.external_id)
+
+                self.assertEqual(mock.call_args[1]['json']['messages'][0]['text'],
+                                 "Test message")
+
+                self.assertEqual(mock.call_args[1]['json']['messages'][0]['notifyContentType'],
+                                 "application/json")
+
+                self.assertEqual(mock.call_args[1]['json']['messages'][0]['notifyUrl'],
+                                 'http://%s%s' % (settings.HOSTNAME, reverse('courier.ib',
+                                                                             args=[self.channel.uuid, 'delivered'])))
+
+                self.assertTrue(mock.call_args[1]['json']['messages'][0]['intermediateReport'])
 
                 self.clear_cache()
 
@@ -4563,6 +4632,26 @@ class InfobipTest(TembaTest):
                 self.assertEqual(2, msg.error_count)
                 self.assertTrue(msg.next_attempt)
 
+            Msg.objects.all().delete()
+            msg = joe.send("Test message", self.admin, trigger_send=False)[0]
+
+            with patch('requests.post') as mock:
+                mock.return_value = MockResponse(200, json.dumps(dict(messages=[
+                    {'status': {'groupId': 2, 'description': "Request was rejected"}, 'messageid': 12}])))
+
+                # manually send it off
+                Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
+
+                # check the status of the message is now sent
+                msg.refresh_from_db()
+                self.assertEqual(ERRORED, msg.status)
+                self.assertEqual(1, msg.error_count)
+                self.assertTrue(msg.next_attempt)
+
+                self.assertEqual(msg.channel_logs.all().count(), 1)
+                self.assertEqual(msg.channel_logs.all().first().description,
+                                 "Received error status: Request was rejected")
+
         finally:
             settings.SEND_MESSAGES = False
 
@@ -4574,7 +4663,7 @@ class InfobipTest(TembaTest):
             settings.SEND_MESSAGES = True
 
             with patch('requests.post') as mock:
-                mock.return_value = MockResponse(200, json.dumps(dict(messages=[{'status': {'id': 0}, 'messageid': 12}])))
+                mock.return_value = MockResponse(200, json.dumps(dict(messages=[{'status': {'groupId': 1}}])))
 
                 # manually send it off
                 Channel.send_message(dict_to_struct('MsgStruct', msg.as_task_json()))
@@ -4583,10 +4672,19 @@ class InfobipTest(TembaTest):
                 msg.refresh_from_db()
                 self.assertEqual(SENT, msg.status)
                 self.assertTrue(msg.sent_on)
-                self.assertEqual('12', msg.external_id)
+                self.assertIsNone(msg.external_id)
 
-                self.assertEqual(mock.call_args[1]['json']['text'],
+                self.assertEqual(mock.call_args[1]['json']['messages'][0]['text'],
                                  "Test message\nhttps://example.com/attachments/pic.jpg")
+
+                self.assertEqual(mock.call_args[1]['json']['messages'][0]['notifyContentType'],
+                                 "application/json")
+
+                self.assertEqual(mock.call_args[1]['json']['messages'][0]['notifyUrl'],
+                                 'http://%s%s' % (settings.HOSTNAME, reverse('courier.ib',
+                                                                             args=[self.channel.uuid, 'delivered'])))
+
+                self.assertTrue(mock.call_args[1]['json']['messages'][0]['intermediateReport'])
 
                 self.clear_cache()
         finally: